ai-risks-htw-2025-main/toy_qa/requirements.txt
any-agent[all,a2a]==0.25.0
PyPDF2
python-dotenv
fire